+++ exampleApp/Main.hs
@@ -35,15 +35,24 @@ featureDummy :: Definition
    -> Feature "dummy" Count)
 featureDummy = define $ pure 5
 
+-- | Lift a subject's events in a feature
+anotherDummy :: Bool 
+  -> Definition
+   ( Feature "allEvents" (Events Day)
+   -> Feature "another" Bool)
+anotherDummy x = define $ const x
+
 -- | Include the subject if she has an enrollment interval concurring with index.
 critTrue :: Definition
    ( Feature "allEvents" (Events Day)
    -> Feature "dummy" Status)
-critTrue = define $ pure Include 
+critTrue = define $ pure Include
 
 instance HasAttributes "dummy" Count where
-  getAttributes _ = emptyAttributes  
+  getAttributes _ = emptyAttributes
 
+instance HasAttributes "another" Bool where
+  getAttributes _ = emptyAttributes
 {-------------------------------------------------------------------------------
   Cohort Specifications and evaluation
 -------------------------------------------------------------------------------}
@@ -58,8 +67,10 @@ makeCriteriaRunner events =
 -- | Make a function that runs the features for a calendar index
 makeFeatureRunner ::
        Events Day
-    -> Featureset 
-makeFeatureRunner events = featureset [packFeature ( eval featureDummy ef )]
+    -> Featureset
+makeFeatureRunner events = featureset 
+    ( packFeature ( eval featureDummy ef ) :|
+      [packFeature ( eval (anotherDummy True ) ef)])
     where ef  = featureEvents events
 
 -- | Make a cohort specification for each calendar time
@@ -67,4 +78,4 @@ cohortSpecs :: [CohortSpec (Events Day) Featureset]
 cohortSpecs = [specifyCohort makeCriteriaRunner makeFeatureRunner]
 
 main :: IO ()
-main = makeCohortApp "testCohort" "v0.1.0" cohortSpecs
+main = makeCohortApp "testCohort" "v0.1.0" colWise cohortSpecs