+++ src/FeatureCompose/Featureable.hs
@@ -5,9 +5,10 @@ Copyright   : (c) NoviSci, Inc 2020
 License     : BSD3
 Maintainer  : bsaul@novisci.com
 -}
-
+{-# OPTIONS_HADDOCK hide #-}
 {-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
 
 module FeatureCompose.Featureable (
     Featureable(..)
@@ -18,7 +19,7 @@ module FeatureCompose.Featureable (
 import FeatureCompose                 ( Feature
                                       , makeFeature
                                       )
-import FeatureCompose.Aeson           ()
+import FeatureCompose.FeatureOutput   ( ShapeOutput(..), OutputShape )
 import FeatureCompose.Attributes      ( HasAttributes(..), Attributes )
 import GHC.TypeLits                   ( KnownSymbol )
 import Data.Aeson                     ( ToJSON(toJSON) )
@@ -27,11 +28,11 @@ import Data.Typeable                  ( Typeable )
 {- | Existential type to hold features, which allows for Features to be put
 into a heterogeneous list.
 -}
-data Featureable = forall d e. (Show d, ToJSON d) => MkFeatureable d  Attributes 
+data Featureable = forall d . (Show d, ToJSON d, ShapeOutput d) => MkFeatureable d  Attributes 
 
 {- | Pack a feature into a @Featurable@. -}
 packFeature ::
-  ( KnownSymbol n, Show d, ToJSON d, Typeable d, HasAttributes n d  ) =>
+  ( KnownSymbol n, Show d, ToJSON d, Typeable d, HasAttributes n d) =>
   Feature n d -> Featureable
 packFeature x = MkFeatureable x (getAttributes x)
 
@@ -40,6 +41,13 @@ instance Show Featureable where
 
 instance ToJSON Featureable where
   toJSON (MkFeatureable x _ ) = toJSON x
+
+instance ShapeOutput Featureable where
+  dataOnly (MkFeatureable x _ ) = dataOnly x
+  nameOnly (MkFeatureable x _ ) = nameOnly x 
+  attrOnly (MkFeatureable x _ ) = attrOnly x 
+  nameData (MkFeatureable x _ ) = nameData x 
+  nameAttr (MkFeatureable x _ ) = nameAttr x 
 
 -- | Get the @Attributes@ from a @Featureable@.
 getFeatureableAttrs :: Featureable -> Attributes