+++ test/Hasklepias/Cohort/CoreSpec.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE DataKinds #-}
 
-module Hasklepias.CohortSpec (
+module Hasklepias.Cohort.CoreSpec (
   spec
  ) where
 
@@ -56,8 +56,9 @@ testCohort = specifyCohort buildCriteria buildFeatures
 testOut :: Cohort Features
 testOut = MkCohort
   ( Just $ MkAttritionInfo $ (ExcludedBy (1, "feat2"), 1) :| [ (Included, 1) ]
-  , [MkObsUnit ("2", ( makeFeature (featureDataR False)
-                     , makeFeature (featureDataR 56))) ])
+  , MkCohortData [MkObsUnit "2" 
+                  ( makeFeature (featureDataR False)
+                  , makeFeature (featureDataR 56)) ])
 
 -- evalCohort testCohort testPopulation
 spec :: Spec